refactor: Rename tasks to task
