+++ crates/scheduler/src/channel/manual.rs
@@ -22,7 +22,7 @@ use tokio::task::JoinHandle;
 use tokio_util::sync::CancellationToken;
 
 use crate::Result;
-use crate::tasks::TaskRequest;
+use crate::task::TaskRequest;
 
 use super::TaskRequestChannel;
